Extract agent tool argument normalizer
