+++ docs/DECISIONS.md
@@ -305,3 +305,4 @@
 - Slash-command dispatch planning lives in `WorkspaceSlashCommandDispatchPlanner`. Parsed `SlashCommand` cases map to typed `WorkspaceSlashCommandDispatchAction` values outside `WorkspaceModel`; the model applies those actions because it owns state, persistence, tool execution, and top-bar sequencing.
 - Native command SF Symbol mapping lives in `QuillCodeCommandIconCatalog`. Sidebar presentation may keep sidebar-specific labels, HTML icon tokens, and deliberate compact-menu overrides, but command palette rows and sidebar rows should not grow separate native icon switches for the same command IDs.
 - `FileSecretStore` is the fallback/dev implementation of `QuillSecretStore`, not the final platform secret backend. It must keep the secrets directory at `0700`, written secret files at `0600`, and sanitize secret keys into a single filename component. App and agent code should continue depending only on `QuillSecretStore` so Keychain, libsecret, or encrypted-file adapters can replace the fallback without changing call sites.
+- TrustedRouter action parsing and tool-argument normalization are separate boundaries. `AgentActionJSONParser` owns extracting a single action object and constructing `AgentAction`; `AgentToolArgumentNormalizer` owns canonical argument keys, tolerated aliases, empty-shell-command repair from explicit nearby prose, and minimum-argument policy. Keep provider transport, JSON-object scanning, prose recovery, and tool schema alias policy in separate files so model-output robustness can improve without turning the parser into a tool registry.
